backend_py/requirement.txt
# Core numerical
numpy>=1.23

# OpenCV (headless for server use)
opencv-python-headless>=4.8

# Environment variables
python-dotenv>=1.0

# FastAPI stack
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5

# Supabase client
supabase>=2.3

# InsightFace + runtime
onnxruntime>=1.16
insightface>=0.7